feat(member): add react-native-screenguard for enhanced screenshot detection and update ID card component
