+++ package.json
@@ -97,8 +97,8 @@
 		"react-native-pager-view": "6.8.1",
 		"react-native-reanimated": "~3.17.5",
 		"react-native-safe-area-context": "5.4.0",
+		"react-native-screenguard": "^1.0.9",
 		"react-native-screens": "^4.11.1",
-		"react-native-screenshot-aware": "^1.3.3",
 		"react-native-select-dropdown": "^4.0.1",
 		"react-native-svg": "15.12.0",
 		"react-native-unistyles": "2.32.0",
