+++ src/components/Member/styles.ts
@@ -9,125 +9,6 @@ export const stylesheet = createStyleSheet((theme) => ({
 		paddingBottom: 30
 	},
 	cardWrapper: { marginBottom: 50 },
-	shadow: {
-		shadowColor: theme.colors.text,
-		shadowOffset: { width: 0, height: 0 },
-		shadowOpacity: 0.2,
-		shadowRadius: 10,
-		elevation: 10
-	},
-	idCardContainer: {
-		borderRadius: theme.radius.lg,
-		overflow: 'visible'
-	},
-	idCard: {
-		borderRadius: theme.radius.lg,
-		padding: 24,
-		minHeight: 200,
-		justifyContent: 'space-between',
-		backgroundColor: '#000000'
-	},
-	cardHeader: {
-		flexDirection: 'row',
-		alignItems: 'center',
-		justifyContent: 'space-between',
-		marginBottom: 20
-	},
-	logoContainer: {
-		alignItems: 'center',
-		justifyContent: 'center'
-	},
-	titleContainer: {
-		alignItems: 'flex-end'
-	},
-	logoText: {
-		color: '#ffffff',
-		fontSize: 20,
-		fontWeight: 'bold'
-	},
-	cardTitle: {
-		color: '#00ff33',
-		fontSize: 12,
-		fontWeight: '700',
-		letterSpacing: 1.5,
-		opacity: 0.9
-	},
-	cardContent: {
-		flex: 1,
-		justifyContent: 'center'
-	},
-	nameSection: {
-		marginBottom: 20
-	},
-	nameLabel: {
-		color: theme.colors.neulandGreen,
-		fontSize: 12,
-		fontWeight: '600',
-		letterSpacing: 1,
-		opacity: 0.8,
-		marginBottom: 4,
-		textTransform: 'uppercase'
-	},
-	name: {
-		color: '#ffffff',
-		fontSize: 32,
-		fontWeight: 'bold',
-		letterSpacing: 0.5
-	},
-	usernameSection: {
-		marginBottom: 20
-	},
-	emailSection: {
-		marginBottom: 12
-	},
-	fieldLabel: {
-		color: theme.colors.neulandGreen,
-		fontSize: 12,
-		fontWeight: '600',
-		letterSpacing: 1,
-		opacity: 0.8,
-		marginBottom: 4,
-		textTransform: 'uppercase'
-	},
-	username: {
-		color: '#ffffff',
-		fontSize: 18,
-		fontWeight: '500',
-		opacity: 0.9
-	},
-	email: {
-		color: '#ffffff',
-		fontSize: 18,
-		fontWeight: '500',
-		opacity: 0.9
-	},
-	qrCodeContainer: {
-		backgroundColor: '#ffffff',
-		padding: 10,
-		borderRadius: theme.radius.md,
-		marginTop: 20,
-		alignItems: 'center',
-		justifyContent: 'center',
-		width: 140,
-		height: 140,
-		alignSelf: 'center'
-	},
-	cardFooter: {
-		marginTop: 20
-	},
-	footerLine: {
-		height: 1,
-		backgroundColor: theme.colors.neulandGreen,
-		marginBottom: 12,
-		opacity: 0.8
-	},
-	footerText: {
-		color: theme.colors.neulandGreen,
-		fontSize: 12,
-		fontWeight: '500',
-		opacity: 0.8,
-		textAlign: 'center'
-	},
 	groupList: {
 		padding: 24
 	},
