+++ src/data/licenses.json
@@ -424,16 +424,16 @@
 		"licenseUrl": "https://github.com/th3rdwave/react-native-safe-area-context/raw/master/LICENSE",
 		"parents": "neuland"
 	},
+	"react-native-screenguard@1.0.9": {
+		"licenses": "MIT",
+		"repository": "https://github.com/gbumps/react-native-screenguard",
+		"licenseUrl": "https://github.com/gbumps/react-native-screenguard/raw/master/LICENSE",
+		"parents": "neuland"
+	},
 	"react-native-screens@4.11.1": {
 		"licenses": "MIT",
 		"repository": "https://github.com/software-mansion/react-native-screens",
 		"licenseUrl": "https://github.com/software-mansion/react-native-screens/raw/master/LICENSE",
-		"parents": "neuland"
-	},
-	"react-native-screenshot-aware@1.3.3": {
-		"licenses": "MIT",
-		"repository": "https://github.com/huextrat/react-native-screenshot-aware",
-		"licenseUrl": "https://github.com/huextrat/react-native-screenshot-aware/raw/master/LICENSE",
 		"parents": "neuland"
 	},
 	"react-native-select-dropdown@4.0.1": {
